requirements.txt
discord.py
python-dotenv
aws_lambda_powertools